requirements.txt
openai
langchain
langchain-community
langchain-openai
faiss-cpu
tqdm
python-dotenv
matplotlib
pandas